+++ src/lib/index.js
@@ -5,7 +5,6 @@
 /* eslint-disable no-unused-vars */
 
 // aqui exportaras las funciones que necesites
-// import { feed } from './lib/views/templateFeed.js';
 
 export const myFunction = () => {
   //  console.log('Hola');
@@ -45,16 +44,42 @@ export const registerFirebase = (signupEmail, signupPassword) => {
     });
 };
 
+// Obtén la información de perfil de un usuario
+// Obtener el perfil de un usuario
+const userInformation = () => {
+  const user = firebase.auth().currentUser;
+  console.log(user, 'entre a la funcion');
+  let name; let email; let photoUrl; let uid; let
+    emailVerified;
+
+  if (user != null) {
+    name = user.displayName;
+    email = user.email;
+    photoUrl = user.photoURL;
+    emailVerified = user.emailVerified;
+    uid = user.uid; // The user's ID, unique to the Firebase project. Do NOT use
+    // this value to authenticate with your backend server, if
+    // you have one. Use User.getToken() instead.
+  }
+
+  if (user != null) {
+    user.providerData.forEach((profile) => {
+      console.log(`Sign-in provider: ${profile.providerId}`);
+      console.log(`  Provider-specific UID: ${profile.uid}`);
+      console.log(`  Name: ${profile.displayName}`);
+      console.log(`  Email: ${profile.email}`);
+      console.log(`  Photo URL: ${profile.photoURL}`);
+    });
+  }
+};
+
 // Signin
 export const signInFirebase = (email, password) => {
   auth
     .signInWithEmailAndPassword(email, password)
     .then(() => {
-      // let url = window.location.hash;
-      // conlsole.log(url);
+      userInformation();
       window.location.hash = ('#/feed');
-      // conlsole.log('shao');
-
     })
     .catch((error) => {
       const errorCode = error.code;
@@ -66,47 +91,15 @@ export const signInFirebase = (email, password) => {
     });
 };
 
-/* const authFeed = () => {
-  showTemplate(hash);
-  if (signInFirebase === 'false' || signInFirebase === '') {
-    alert('Completa tus datos correctamente para ingresar');
-  } else {
-    signInFirebase(email, password)
-      .then(() => {
-        changeRoute('#/feed');
-      });
-  }
-};
-authFeed(); */
-
-// Obtener el perfil de un usuario
-const user = firebase.auth().currentUser;
-console.log(user);
-var name, email, photoUrl, uid, emailVerified;
-
-if (user != null) {
-  name = user.displayName;
-  email = user.email;
-  photoUrl = user.photoURL;
-  emailVerified = user.emailVerified;
-  uid = user.uid; // The user's ID, unique to the Firebase project. Do NOT use
-  // this value to authenticate with your backend server, if
-  // you have one. Use User.getToken() instead.
-}
-
-// Obtén la información de perfil de un usuario
-
-if (user != null) {
-  user.providerData.forEach((profile) => {
-    console.log(`Sign-in provider: ${profile.providerId}`);
-    console.log(`  Provider-specific UID: ${profile.uid}`);
-    console.log(`  Name: ${profile.displayName}`);
-    console.log(`  Email: ${profile.email}`);
-    console.log(`  Photo URL: ${profile.photoURL}`);
-  });
-}
 // Singin with google
 export const googleProvider = () => {
   const provider = new firebase.auth.GoogleAuthProvider();
-  return firebase.auth().signInWithPopup(provider);
+  firebase.auth().signInWithPopup(provider).then((result) => {
+    userInformation();
+    window.location.hash = ('#/feed');
+    // This gives you a Google Access Token.
+    const token = result.credential.accessToken;
+    // The signed-in user info.
+    const user = result.user;
+  });
 };
