+++ src/style.css
@@ -9,7 +9,7 @@ body {
     box-sizing: border-box;
 }
 header{
-    padding: 20px;
+    padding: 10px;
     padding-bottom: 0;
     border-bottom: 1px solid rgb(0, 0, 0);
     display: flex;
@@ -173,6 +173,7 @@ h5 {
     color: rgb(0, 0, 0);
 }
 .Container-camera{
+    margin-top: 20px;
     width: 100%;
     height: 100%;
     background-color: #E5E5E5;
